requirements.txt
vnstock>=3.0.9
yfinance>=0.2.40
pandas>=2.0.0
pandas_ta>=0.4.0
numpy>=1.24.0
python-dateutil>=2.8.2
